Move the workload e2e tests to it's own package
