+++ test/e2e/BUILD
@@ -28,6 +28,7 @@ go_test(
         "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/storage:go_default_library",
+        "//test/e2e/workload:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
@@ -50,12 +51,8 @@ go_library(
         "cadvisor.go",
         "certificates.go",
         "cluster_upgrade.go",
-        "cronjob.go",
         "custom_resource_definition.go",
-        "daemon_restart.go",
-        "daemon_set.go",
         "dashboard.go",
-        "deployment.go",
         "disruption.go",
         "dns.go",
         "dns_common.go",
@@ -74,7 +71,6 @@ go_library(
         "gke_node_pools.go",
         "ha_master.go",
         "ingress.go",
-        "job.go",
         "kibana_logging.go",
         "kube_proxy.go",
         "kubectl.go",
@@ -97,9 +93,7 @@ go_library(
         "portforward.go",
         "pre_stop.go",
         "proxy.go",
-        "rc.go",
         "reboot.go",
-        "replica_set.go",
         "resize_nodes.go",
         "resource_quota.go",
         "restart.go",
@@ -110,7 +104,6 @@ go_library(
         "serviceloadbalancers.go",
         "ssh.go",
         "stackdriver_monitoring.go",
-        "statefulset.go",
         "third-party.go",
         "ubernetes_lite.go",
         "util_iperf.go",
@@ -120,25 +113,17 @@ go_library(
         "//pkg/api:go_default_library",
         "//pkg/api/v1/pod:go_default_library",
         "//pkg/api/v1/service:go_default_library",
-        "//pkg/apis/batch:go_default_library",
-        "//pkg/apis/extensions:go_default_library",
+        "//pkg/apis/batch/v2alpha1:go_default_library",
         "//pkg/apis/networking:go_default_library",
         "//pkg/client/clientset_generated/clientset:go_default_library",
         "//pkg/client/clientset_generated/clientset/typed/certificates/v1beta1:go_default_library",
-        "//pkg/client/clientset_generated/clientset/typed/extensions/v1beta1:go_default_library",
         "//pkg/client/clientset_generated/internalclientset:go_default_library",
         "//pkg/cloudprovider:go_default_library",
         "//pkg/cloudprovider/providers/azure:go_default_library",
         "//pkg/cloudprovider/providers/gce:go_default_library",
         "//pkg/controller:go_default_library",
-        "//pkg/controller/daemon:go_default_library",
-        "//pkg/controller/deployment/util:go_default_library",
         "//pkg/controller/endpoint:go_default_library",
-        "//pkg/controller/job:go_default_library",
         "//pkg/controller/node:go_default_library",
-        "//pkg/controller/replicaset:go_default_library",
-        "//pkg/controller/replication:go_default_library",
-        "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubelet/apis:go_default_library",
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
@@ -146,12 +131,10 @@ go_library(
         "//pkg/master/ports:go_default_library",
         "//pkg/metrics:go_default_library",
         "//pkg/quota/evaluator/core:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/util/exec:go_default_library",
         "//pkg/util/logs:go_default_library",
         "//pkg/util/version:go_default_library",
         "//plugin/pkg/admission/serviceaccount:go_default_library",
-        "//plugin/pkg/scheduler/schedulercache:go_default_library",
         "//test/e2e/chaosmonkey:go_default_library",
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
@@ -161,11 +144,11 @@ go_library(
         "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/upgrades:go_default_library",
+        "//test/e2e/workload:go_default_library",
         "//test/e2e_federation:go_default_library",
         "//test/images/net/nat:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/junit:go_default_library",
-        "//vendor/github.com/davecgh/go-spew/spew:go_default_library",
         "//vendor/github.com/elazarl/goproxy:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
@@ -179,7 +162,6 @@ go_library(
         "//vendor/golang.org/x/net/websocket:go_default_library",
         "//vendor/golang.org/x/oauth2/google:go_default_library",
         "//vendor/google.golang.org/api/monitoring/v3:go_default_library",
-        "//vendor/k8s.io/api/apps/v1beta1:go_default_library",
         "//vendor/k8s.io/api/batch/v1:go_default_library",
         "//vendor/k8s.io/api/batch/v2alpha1:go_default_library",
         "//vendor/k8s.io/api/certificates/v1beta1:go_default_library",
@@ -191,7 +173,6 @@ go_library(
         "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/test/integration/testserver:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -254,6 +235,7 @@ filegroup(
         "//test/e2e/storage:all-srcs",
         "//test/e2e/testing-manifests:all-srcs",
         "//test/e2e/upgrades:all-srcs",
+        "//test/e2e/workload:all-srcs",
     ],
     tags = ["automanaged"],
 )
