+++ test/e2e/e2e_test.go
@@ -27,6 +27,7 @@ import (
 	_ "k8s.io/kubernetes/test/e2e/perf"
 	_ "k8s.io/kubernetes/test/e2e/scheduling"
 	_ "k8s.io/kubernetes/test/e2e/storage"
+	_ "k8s.io/kubernetes/test/e2e/workload"
 )
 
 func init() {
