+++ test/e2e/etcd_failure.go
@@ -24,6 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	podutil "k8s.io/kubernetes/pkg/api/v1/pod"
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/workload"
 	testutils "k8s.io/kubernetes/test/utils"
 
 	. "github.com/onsi/ginkgo"
@@ -73,7 +74,7 @@ func etcdFailTest(f *framework.Framework, failCommand, fixCommand string) {
 
 	checkExistingRCRecovers(f)
 
-	testReplicationControllerServeImageOrFail(f, "basic", framework.ServeHostnameImage)
+	workload.TestReplicationControllerServeImageOrFail(f, "basic", framework.ServeHostnameImage)
 }
 
 // For this duration, etcd will be failed by executing a failCommand on the master.
