+++ test/e2e/workload/OWNERS
@@ -1,0 +1,7 @@
+approvers:
+- janetkuo
+- nikhiljindal
+- kargakis
+- mfojtik
+reviewers:
+- sig-apps-reviewers
